requirements.txt
yt-dlp
ffmpeg-python
requests
beautifulsoup4
selenium
openai
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
